console: Refactor device general settings
Move redux related code outside of the view component
